fix/fixed_auction_info_deserealization
Fixed reserved_slots type

+++ types/bid.go
@@ -28,7 +28,7 @@ type ValidatorBid struct {
 	// Vesting schedule for a genesis validator. `None` if non-genesis validator.
 	VestingSchedule *VestingSchedule `json:"vesting_schedule"`
 	// Number of slots reserved for specific delegators
-	ReservedSlots uint8 `json:"reserved_slots"`
+	ReservedSlots uint32 `json:"reserved_slots"`
 }
 
 // Bid An entry in the validator map.
